allow test running by file name and line number
instead of passing an ns-re, now you can pass a regex (works like
before), or a string. If a string is passed, and it is in the form
`file-name:line-number` (e.g. test/bolth_test.clj:18), then the lowest
test var(s) whole line number is above or on the line given is run.
If the string given is a valid clojure filename (with no line number),
then just tests from that file are run. If the string given isn't either
a filename with a line number or a filename, it's treated as a regex.